refactor(compiler): reduce intent to entirely deterministic
As the _big programming guy_ ayaz hafiz says,

```
the clearest intent is no intent at all
```

Which as a philosophical statement is obvious. We should aspire to do
the same thing in the software we engineer - to be as clear as possible
in the code we right.

This commit makes the operation of the `mediterranean` C compiler as
clear and deterministic is it possibly could be. By eliding all
business logic, the intention of the compiler is unambigious.

Some may say this is not code. Some may say this makes the compiler
non-functional. They are not incorrect. But software, like all art, is
in the eye of the engineer. Perhaps Haskell has not caught up yet, but
this is the ultimate abstraction we could hope to have in the field of
Computer Science and Computer Programming.